Revert "Refacto/restucturation"
