+++ .gitignore
@@ -51,7 +51,6 @@ coverage.xml
 .hypothesis/
 .pytest_cache/
 cover/
-.vscode/
 
 # Translations
 *.mo
