+++ pytest.ini
@@ -7,5 +7,3 @@ addopts = -v --tb=short --strict-markers
 markers =
     slow: marks tests as slow
     integration: marks tests as integration tests
-    unit: marks tests as unit tests
-    cli: marks tests as CLI tests
