--- src/classes/errors.py
@@ -1,10 +1,0 @@
-class TaskValidationError(Exception):
-    """Exception personnalisée pour les erreurs de validation"""
-
-    pass
-
-
-class TaskNotFoundError(Exception):
-    """Exception personnalisée pour tâche non trouvée"""
-
-    pass
