+++ tasks.json
@@ -1,14 +1,21 @@
 [
   {
     "id": 1,
-    "title": "Nouveau titre",
-    "description": "Nouvelle description",
-    "status": "DONE",
+    "title": "Première tâche",
+    "description": "Description de la première tâche",
+    "status": "TODO",
     "created_at": "2024-01-01T10:00:00"
   },
   {
+    "id": 2,
+    "title": "Deuxième tâche",
+    "description": "Description de la deuxième tâche",
+    "status": "DONE",
+    "created_at": "2024-01-02T15:00:00"
+  },
+  {
     "id": 3,
-    "title": "Nouveau titre test",
+    "title": "Persistée",
     "description": "",
     "status": "TODO",
     "created_at": "2025-06-30T15:18:13"
@@ -26,12 +33,5 @@
     "description": "Doit être faite rapidement",
     "status": "TODO",
     "created_at": "2025-07-01T14:58:28"
-  },
-  {
-    "id": 6,
-    "title": "Test création",
-    "description": "test desrcription nouvelle",
-    "status": "TODO",
-    "created_at": "2025-07-02T11:22:41"
   }
 ]
